Watch for site changes & rebuild when serving
This makes it so that the server watches for file changes to improve the
development experience.

Maybe there would be value in not watching and building at some point
with the `soy server` command, but this works for most cases for now.

Uses the listen gem because it's mature and working well from my initial
implementation.