Add support for `soy server` to serve bare HTML
This makes it possible to have a file in the build named `about.html`
and request it at `http://localhost:9292/about`, which is standard for static
site generators and hosts. `http://localhost:9292/about.html` will still
work as well.

Netlify and GitHub Pages support this as well for their hosting, so
that's also nice.

It might be necessary to support build output config that generates
something like `build/about/index.html` for the `about.html` file, but I
can cross that bridge when I get there.