[experiment] define routes & add helpers for rendering them
This is an exploration into seeing what it'd be like to define routes
and call out to them via ways other than hardcoding links via the anchor
href. But idk if this really gets us much since the generate URLs will
be based on filenames and such. Maybe once models are introduced this
will be more useful...