Refactor exe/soy to use a class
This makes the data structure for managing commands a lot nicer and
easier to maintain. It also DRYs up the code in a lot of ways in that
the help docs are automatically generated.

A major motivator here is to make it easier to unit test the various
commands and remove the alias system specs since they're the slowest
specs in the whole suite and should be used sparingly.