Add support for --port opt in `soy server`
Allows for avoiding port conflicts and running multiple soy servers on a
machine at one time. Uses OptionParser because it's legit amazing and
serves options for these commands really well.

Eventually it'd be great to be able to configure the port in another
way, like via site config, but this works for now.