Add support for nested content
This allows for more complex organization of content by supporting
loading and outputting build content within directories.

Also, a sort of hidden change in the code here is that the build dir
gets cleaned before each build to ensure it's clean and there are
minimal conflicts. This may be a bit too much a sledgehammer approach
for the long-term, but it works for now.